Update 26/02/2022
Added new test seed document, with a graph with only 1 component to ensure the test of "diameter()" function.
Changed the method of testing "how_many_degrees()" to make it faster, now it doesn't call de method "how_many_degrees()" for each pair of nodes, it calls the method "floyd_algorithm()" that returns the shortest path between each pair of nodes.
Automated the condition to enter the test with lastfm-graph and automated the print of results.
Resolved the problem with graph with more than one component, now returns "None" if there is more than one component.